requirements.txt
python-dotenv
google-genai
Flask
